+++ verticapy/io/flex.py
@@ -15,7 +15,8 @@ See the  License for the specific  language governing
 permissions and limitations under the License.
 """
 
-from verticapy.utils._toolbox import executeSQL, quote_ident
+from verticapy.utils._toolbox import executeSQL
+from verticapy.io.sql.utils._format import quote_ident
 from verticapy.core.str_sql import str_sql
 from verticapy.errors import ParameterError
 from verticapy.utils._decorators import save_verticapy_logs
