+++ verticapy/io/parsers/csv.py
@@ -28,6 +28,7 @@ from verticapy.utils._decorators import save_verticapy_logs
 from verticapy.utils._toolbox import *
 from verticapy.errors import ExtensionError, ParameterError, MissingRelation
 from ..flex import compute_flextable_keys
+from verticapy.io.sql.utils._format import format_schema_table, clean_query
 
 
 def guess_sep(file_str: str):
