+++ verticapy/io/parsers/json.py
@@ -23,6 +23,7 @@ from verticapy.utils._decorators import save_verticapy_logs
 from verticapy.utils._toolbox import *
 from verticapy.errors import ExtensionError, ParameterError, MissingRelation
 from ..flex import compute_flextable_keys
+from verticapy.io.sql.utils._format import quote_ident, format_schema_table, clean_query
 
 
 def pjson(path: str, ingest_local: bool = True):
