+++ verticapy/io/sql/create.py
@@ -34,7 +34,8 @@ Returns
 bool
     True if the schema was successfully created, False otherwise.
     """
-    from verticapy.utils._toolbox import executeSQL, quote_ident
+    from verticapy.utils._toolbox import executeSQL
+    from verticapy.io.sql.utils._format import quote_ident
 
     try:
         executeSQL(f"CREATE SCHEMA {schema};", title="Creating the new schema.")
