+++ verticapy/io/sql/drop.py
@@ -16,6 +16,12 @@ permissions and limitations under the License.
 """
 from typing import Literal
 
+from verticapy.io.sql.utils._format import (
+    quote_ident,
+    schema_relation,
+    format_schema_table,
+)
+
 
 def drop(
     name: str = "",
@@ -51,7 +57,7 @@ Returns
 bool
     True if the relation was dropped, False otherwise.
     """
-    from verticapy.utils._toolbox import executeSQL, quote_ident, schema_relation
+    from verticapy.utils._toolbox import executeSQL
 
     if "relation_type" in kwds and method == "auto":
         method = kwds["relation_type"]
