+++ verticapy/io/sql/insert.py
@@ -17,6 +17,7 @@ permissions and limitations under the License.
 import warnings, sys, time
 from verticapy.utils._decorators import save_verticapy_logs
 from verticapy.errors import MissingRelation
+from verticapy.io.sql.utils._format import format_schema_table, clean_query, quote_ident
 
 
 @save_verticapy_logs
@@ -59,12 +60,7 @@ See Also
 pandas_to_vertica : Ingests a pandas DataFrame into the Vertica database.
     """
     import verticapy as vp
-    from verticapy.utils._toolbox import (
-        executeSQL,
-        format_schema_table,
-        quote_ident,
-        clean_query,
-    )
+    from verticapy.utils._toolbox import executeSQL
 
     if not (schema):
         schema = vp.OPTIONS["temp_schema"]
