+++ verticapy/machine_learning/vertica/tsa.py
@@ -34,6 +34,7 @@ from verticapy.learn.linear_model import LinearRegression
 from verticapy import vDataFrame, save_verticapy_logs
 from verticapy.plotting._colors import gen_colors
 from verticapy.learn.tools import *
+from verticapy.io.sql.utils._format import quote_ident, schema_relation
 
 # Other Python Modules
 from dateutil.parser import parse
@@ -1271,7 +1272,6 @@ solver: str, optional
     object
         self
         """
-        # Initialization
         if verticapy.OPTIONS["overwrite_model"]:
             self.drop()
         else:
