+++ verticapy/plotting/_matplotlib/animated.py
@@ -27,7 +27,7 @@ import numpy as np
 
 # VerticaPy Modules
 from verticapy.utilities import *
-from verticapy.utils._toolbox import executeSQL, quote_ident, updated_dict
+from verticapy.utils._toolbox import executeSQL, updated_dict
 from verticapy.errors import ParameterError
 from verticapy.plotting._colors import gen_cmap, gen_colors
