+++ verticapy/plotting/_matplotlib/boxplot.py
@@ -22,7 +22,7 @@ import matplotlib.pyplot as plt
 
 # VerticaPy Modules
 from verticapy.utilities import *
-from verticapy.utils._toolbox import executeSQL, quote_ident
+from verticapy.utils._toolbox import executeSQL
 from verticapy.errors import ParameterError
 from verticapy.plotting._colors import gen_colors
