+++ verticapy/plotting/_matplotlib/core.py
@@ -23,9 +23,10 @@ permissions and limitations under the License.
 import math
 
 # VerticaPy Modules
-from verticapy.utils._toolbox import executeSQL, quote_ident, bin_spatial_to_str
+from verticapy.utils._toolbox import executeSQL, bin_spatial_to_str
 from verticapy.errors import ParameterError
 import verticapy
+from verticapy.io.sql.utils._format import quote_ident
 
 #
 ##
