+++ verticapy/plotting/_matplotlib/spider.py
@@ -22,7 +22,7 @@ import matplotlib.pyplot as plt
 
 # VerticaPy Modules
 from verticapy.utilities import *
-from verticapy.utils._toolbox import executeSQL, quote_ident, updated_dict
+from verticapy.utils._toolbox import executeSQL, updated_dict
 from verticapy.errors import ParameterError
 from verticapy.plotting._colors import gen_colors
