+++ verticapy/stats/_utils.py
@@ -17,6 +17,7 @@ permissions and limitations under the License.
 import itertools
 import numpy as np
 
+
 def all_comb(X: list):
     all_configuration = []
     for r in range(len(X) + 1):
