+++ verticapy/stats/math/math.py
@@ -28,6 +28,7 @@ from verticapy.utilities import *
 from verticapy.utils._toolbox import *
 from verticapy.stats.tools import *
 from verticapy.core.str_sql import str_sql
+from verticapy.io.sql.utils._format import format_magic, clean_query
 
 #
 # Global Variables
