+++ verticapy/tests/vModel/test_xgb_classifier.py
@@ -28,10 +28,11 @@ import xgboost as xgb
 # VerticaPy
 import verticapy
 from verticapy.tests.conftest import get_version
-from verticapy import vDataFrame, drop, set_option, clean_query
+from verticapy import vDataFrame, drop, set_option
 from verticapy.connect import current_cursor
 from verticapy.datasets import load_titanic, load_dataset_cl
 from verticapy.learn.ensemble import XGBoostClassifier
+from verticapy.io.sql.utils._format import clean_query
 
 set_option("print_info", False)
